refactor: remove unnecessary type conversions

+++ drawing/raster_graphic_context.go
@@ -193,7 +193,7 @@ func (rgc *RasterGraphicContext) GetStringBounds(s string) (left, top, right, bo
 // recalc recalculates scale and bounds values from the font size, screen
 // resolution and font metrics, and invalidates the glyph cache.
 func (rgc *RasterGraphicContext) recalc() {
-	rgc.current.Scale = rgc.current.FontSizePoints * float64(rgc.DPI)
+	rgc.current.Scale = rgc.current.FontSizePoints * rgc.DPI
 }
 
 // SetFont sets the font used to draw text.
